[main] Show updates from last saved state when running with no arguments
This commit removes the functionality to preview package updates, as it caused the potentially unwanted side effect of forcing you to
update your system whenever you wanted to rebuild your system configuration. While it may be possible to avoid this by running
"nixos-rebuild --rollback", it does not seem appropriate for a tool of this nature to perform those kinds of actions behind the scenes.

Since the default action of the program is to display updates between the last saved package state, the -f flag has been removed, and
--savestate has been renamed to --save.