remove global dependency handling
This feature was initially added to prevent spam when a major package is updated as it would then show updates to nearly every installed package.
However, it can also be quite useful to see every package that an update was applied to to find regressions.

If this seems to cause too much spam, then a flag to toggle it may be added instead.