replace rusqlite with diesel
This change provides some small performance improvements.

The one downside to this change is that because Diesel does not support setting flags when opening an SQLite connection, the Nix database is opened in read / write mode.
This will be fixed if/when https://github.com/diesel-rs/diesel/pull/1292 is merged.